@@ -14,11 +14,7 @@
 
         <TargetFramework>netstandard2.0</TargetFramework>
 
-        <LangVersion>7.3</LangVersion>
+        <LangVersion>12</LangVersion>
     </PropertyGroup>
-
-    <ItemGroup>
-      <Folder Include="Models" />
-    </ItemGroup>
 
 </Project>
