@@ -1,28 +1,27 @@
-namespace PangLib.PET.Models
+namespace PangLib.PET.Models;
+
+/// <summary>
+/// Positional data structure used inside <see cref="PangLib.PET.Models.Animation"/>, determining position changes
+/// </summary>
+public struct PositionData
 {
     /// <summary>
-    /// Positional data structure used inside <see cref="PangLib.PET.Models.Animation"/>, determining position changes
+    /// Time in seconds
     /// </summary>
-    public struct PositionData
-    {
-        /// <summary>
-        /// Time in seconds
-        /// </summary>
-        public float Time { get; set; }
+    public float Time { get; set; }
         
-        /// <summary>
-        /// X coordinate at <see cref="Time"/>
-        /// </summary>
-        public float X { get; set; }
+    /// <summary>
+    /// X coordinate at <see cref="Time"/>
+    /// </summary>
+    public float X { get; set; }
         
-        /// <summary>
-        /// Y coordinate at <see cref="Time"/>
-        /// </summary>
-        public float Y { get; set; }
+    /// <summary>
+    /// Y coordinate at <see cref="Time"/>
+    /// </summary>
+    public float Y { get; set; }
         
-        /// <summary>
-        /// Z coordinate at <see cref="Time"/>
-        /// </summary>
-        public float Z { get; set; }
-    }
+    /// <summary>
+    /// Z coordinate at <see cref="Time"/>
+    /// </summary>
+    public float Z { get; set; }
 }
