@@ -1,18 +1,17 @@
-namespace PangLib.PET.Models
+namespace PangLib.PET.Models;
+
+/// <summary>
+/// UV mapping structure used inside <see cref="PangLib.PET.Models.PolygonIndex"/>
+/// </summary>
+public struct UVMapping
 {
     /// <summary>
-    /// UV mapping structure used inside <see cref="PangLib.PET.Models.PolygonIndex"/>
+    /// U-axis of the UV mapping
     /// </summary>
-    public struct UVMapping
-    {
-        /// <summary>
-        /// U-axis of the UV mapping
-        /// </summary>
-        public float U { get; set; }
+    public float U { get; set; }
         
-        /// <summary>
-        /// V-axis of the UV mapping
-        /// </summary>
-        public float V { get; set; }
-    }
+    /// <summary>
+    /// V-axis of the UV mapping
+    /// </summary>
+    public float V { get; set; }
 }
