@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
 
-namespace PangLib.SBIN.Models
+namespace PangLib.SBIN.Models;
+
+public struct Texture
 {
-    public struct Texture
-    {
-        public int Width;
-        public int Height;
-        public byte[] TextureData;
-        public List<DDSMap> DDSMaps;
-    }
+    public int Width;
+    public int Height;
+    public byte[] TextureData;
+    public List<DDSMap> DDSMaps;
 }
