@@ -1,15 +1,14 @@
-﻿namespace PangLib.SBIN.Models
+﻿namespace PangLib.SBIN.Models;
+
+public struct Vector2<T>
 {
-    public struct Vector2<T>
-    {
-        public T X;
-        public T Y;
-    }
+    public T X;
+    public T Y;
+}
     
-    public struct Vector3<T>
-    {
-        public T X;
-        public T Y;
-        public T Z;
-    }
+public struct Vector3<T>
+{
+    public T X;
+    public T Y;
+    public T Z;
 }
