@@ -1,36 +1,35 @@
-namespace PangLib.Scripting.PET
+namespace PangLib.Scripting.PET;
+
+/// <summary>
+/// All token types used in <see cref="PETScript"/>
+/// </summary>
+public enum PETTokenType
 {
     /// <summary>
-    /// All token types used in <see cref="PETScript"/>
+    /// An argument in a <see cref="PETCommand"/>
+    ///
+    /// <example>"__facetexture__"</example>
     /// </summary>
-    public enum PETTokenType
-    {
-        /// <summary>
-        /// An argument in a <see cref="PETCommand"/>
-        ///
-        /// <example>"__facetexture__"</example>
-        /// </summary>
-        Argument,
+    Argument,
         
-        /// <summary>
-        /// A closing parenthesis
-        ///
-        /// <example>)</example>
-        /// </summary>
-        CloseParenthesis,
+    /// <summary>
+    /// A closing parenthesis
+    ///
+    /// <example>)</example>
+    /// </summary>
+    CloseParenthesis,
         
-        /// <summary>
-        /// Name of a <see cref="PETCommand"/>
-        ///
-        /// <example>*ptex</example>
-        /// </summary>
-        CommandName,
+    /// <summary>
+    /// Name of a <see cref="PETCommand"/>
+    ///
+    /// <example>*ptex</example>
+    /// </summary>
+    CommandName,
         
-        /// <summary>
-        /// A opening parenthesis
-        ///
-        /// <example>(</example>
-        /// </summary>
-        OpenParenthesis
-    }
+    /// <summary>
+    /// A opening parenthesis
+    ///
+    /// <example>(</example>
+    /// </summary>
+    OpenParenthesis
 }
