@@ -1,10 +1,9 @@
-namespace PangLib.Scripting
+namespace PangLib.Scripting;
+
+public class TokenMatch<T> where T : new()
 {
-    public class TokenMatch<T> where T : new()
-    {
-        public bool IsMatch { get; set; }
-        public T TokenType { get; set; }
-        public string Value { get; set; }
-        public string RemainingText { get; set; }
-    }
+    public bool IsMatch { get; set; }
+    public T TokenType { get; set; }
+    public string Value { get; set; }
+    public string RemainingText { get; set; }
 }
