@@ -5,59 +5,58 @@ using System.Text;
 using System.Xml;
 using PangLib.Utilities.Cryptography;
 
-namespace PangLib.UpdateList
+namespace PangLib.UpdateList;
+
+/// <summary>
+/// Main UpdateList file class
+/// </summary>
+public class UpdateList
 {
+    public string Document;
+
+    private uint[] DecryptionKey;
+    private string FilePath;
+
     /// <summary>
-    /// Main UpdateList file class
+    /// Constructor for the UpdateList class
     /// </summary>
-    public class UpdateList
+    /// <param name="filePath">Path to the updatelist file</param>
+    public UpdateList(string filePath)
     {
-        public string Document;
+        FilePath = filePath;
+    }
 
-        private uint[] DecryptionKey;
-        private string FilePath;
-
-        /// <summary>
-        /// Constructor for the UpdateList class
-        /// </summary>
-        /// <param name="filePath">Path to the updatelist file</param>
-        public UpdateList(string filePath)
+    /// <summary>
+    /// Decrypts and parses the updatelist file
+    ///
+    /// Saves the parsed updatelist in the Document instance attribute
+    /// </summary>
+    public void Parse()
+    {
+        if (DecryptionKey == null)
         {
-            FilePath = filePath;
+            throw new NullReferenceException("DecryptionKey needs to be set using the SetDecryptionKey instance method!");
         }
 
-        /// <summary>
-        /// Decrypts and parses the updatelist file
-        ///
-        /// Saves the parsed updatelist in the Document instance attribute
-        /// </summary>
-        public void Parse()
+        Span<byte> updateList = File.ReadAllBytes(FilePath);
+
+        for (int i = 0; i < updateList.Length; i += 8)
         {
-            if (DecryptionKey == null)
-            {
-                throw new NullReferenceException("DecryptionKey needs to be set using the SetDecryptionKey instance method!");
-            }
-
-            Span<byte> updateList = File.ReadAllBytes(FilePath);
-
-            for (int i = 0; i < updateList.Length; i += 8)
-            {
-                Span<byte> chunk = updateList.Slice(i, 8);
-                Span<uint> decrypted = XTEA.Decipher(16, MemoryMarshal.Cast<byte, uint>(chunk).ToArray(), DecryptionKey);
-                Span<byte> resource = MemoryMarshal.AsBytes(decrypted);
-                resource.CopyTo(chunk);
-            }
-
-            Document = Encoding.UTF8.GetString(updateList.ToArray());
+            Span<byte> chunk = updateList.Slice(i, 8);
+            Span<uint> decrypted = XTEA.Decipher(16, MemoryMarshal.Cast<byte, uint>(chunk).ToArray(), DecryptionKey);
+            Span<byte> resource = MemoryMarshal.AsBytes(decrypted);
+            resource.CopyTo(chunk);
         }
 
-        /// <summary>
-        /// Sets the decryption key for the updatelist decryption
-        /// </summary>
-        /// <param name="key">Decryption key</param>
-        public void SetDecryptionKey(uint[] key)
-        {
-            DecryptionKey = key;
-        }
+        Document = Encoding.UTF8.GetString(updateList.ToArray());
+    }
+
+    /// <summary>
+    /// Sets the decryption key for the updatelist decryption
+    /// </summary>
+    /// <param name="key">Decryption key</param>
+    public void SetDecryptionKey(uint[] key)
+    {
+        DecryptionKey = key;
     }
 }
