@@ -1,18 +1,17 @@
-﻿namespace PangLib.WEP.Models
+﻿namespace PangLib.WEP.Models;
+
+/// <summary>
+/// Extra values included in most models
+/// </summary>
+public struct ExtraValues
 {
     /// <summary>
-    /// Extra values included in most models
+    /// GUID of the model
     /// </summary>
-    public struct ExtraValues
-    {
-        /// <summary>
-        /// GUID of the model
-        /// </summary>
-        public int GUID { get; set; }
+    public int GUID { get; set; }
         
-        /// <summary>
-        /// Unknown number value
-        /// </summary>
-        public int Unknown { get; set; }
-    }
+    /// <summary>
+    /// Unknown number value
+    /// </summary>
+    public int Unknown { get; set; }
 }
