@@ -1,25 +1,24 @@
-﻿namespace PangLib.WEP.Models
+﻿namespace PangLib.WEP.Models;
+
+/// <summary>
+/// Base vector type
+/// </summary>
+/// <typeparam name="T">Number type</typeparam>
+public struct Vector3<T>
 {
     /// <summary>
-    /// Base vector type
+    /// X coordinate value
     /// </summary>
-    /// <typeparam name="T">Number type</typeparam>
-    public struct Vector3<T>
-    {
-        /// <summary>
-        /// X coordinate value
-        /// </summary>
-        public T X { get; set; }
+    public T X { get; set; }
         
-        /// <summary>
-        /// Y coordinate value
-        /// </summary>
-        public T Y { get; set; }
+    /// <summary>
+    /// Y coordinate value
+    /// </summary>
+    public T Y { get; set; }
         
-        /// <summary>
-        /// Z coordinate value
-        /// </summary>
+    /// <summary>
+    /// Z coordinate value
+    /// </summary>
         
-        public T Z { get; set; }
-    }
+    public T Z { get; set; }
 }
